+++ CHANGES.md
@@ -5,6 +5,7 @@
 - ENH: Add `set.get_mosaics()` to retrieve all the mosaics contained in any set
 - FIX: Don't except ImportErrors in `__init__` and therefore look for version internally in `__meta__` file
 - FIX: Fix band list type when writing a stack on disk
+- FIX: Fix nodata management when stacking a mosaic
 
 ## 0.4.1 (2025-01-07)